Refactored Opmodes and some filenames
