+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/freightfrenzy/auto/other/RedLeft.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.opmodes.freightfrenzy.auto;
+package org.firstinspires.ftc.teamcode.opmodes.freightfrenzy.auto.other;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.geometry.Vector2d;
@@ -29,9 +29,10 @@ import static org.firstinspires.ftc.teamcode.robot.subsystems.drivetrain.DriveCo
 
 import java.util.Locale;
 import java.util.Objects;
+@Disabled
 @Autonomous(preselectTeleOp = "Red TeleOp")
-public class RedRight extends BaseOpMode {
-    Pose2d startPose = new Pose2d(9, -63, Math.toRadians(90));
+public class RedLeft extends BaseOpMode {
+    Pose2d startPose = new Pose2d(-39, -63, Math.toRadians(90));
 
     @Override
     public void preRunLoop() {
