+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/freightfrenzy/auto/other/SharedBlue.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.opmodes.freightfrenzy.auto;/*
+package org.firstinspires.ftc.teamcode.opmodes.freightfrenzy.auto.other;/*
 package org.firstinspires.ftc.teamcode.opmodes.freightfrenzy.auto;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
@@ -13,11 +13,10 @@ import org.firstinspires.ftc.teamcode.robot.util.PositionUtil;
 
 import java.util.Locale;
 import java.util.Objects;
-
 @Disabled
 @Autonomous(preselectTeleOp = "TeleBasic")
-public class AllianceRed extends BaseOpMode {
-    Pose2d startPose = new Pose2d(-34, -61, Math.toRadians(90));
+public class SharedBlue extends BaseOpMode {
+    Pose2d startPose = new Pose2d(10, 61, Math.toRadians(-90));
 
     @Override
     public void pre_setup() {
@@ -36,13 +35,20 @@ public class AllianceRed extends BaseOpMode {
             robot.drivetrain.setPoseEstimate(startPose);
         }
 
+        //robot.jankArm.setAngle(Math.toRadians(200));
+        robot.update();
+
         // Trajectory to get the robot into the shared thing on blue
-        TrajectorySequence toSharedRed = robot.drivetrain.trajectorySequenceBuilder(startPose)
-                .splineTo(new Vector2d(-50, -50), Math.toRadians(135))
-                .splineTo(new Vector2d(-59, -35), Math.toRadians(90))
+        TrajectorySequence toSharedBlue = robot.drivetrain.trajectorySequenceBuilder(startPose)
+                .splineTo(new Vector2d(10, 46), 0)
+                .splineTo(new Vector2d(50, 46), 0)
+                .addDisplacementMarker(() -> {
+                    // Set it to the max (it'll get clipped)
+                    //robot.jankArm.setAngle(10);
+                })
                 .build();
 
-        robot.drivetrain.followTrajectorySequenceAsync(toSharedRed);
+        robot.drivetrain.followTrajectorySequenceAsync(toSharedBlue);
     }
 
     */
