+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/abstracts/BaseOpMode.kt
@@ -20,7 +20,7 @@ SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 package org.firstinspires.ftc.teamcode.robot.abstracts
 
-import org.firstinspires.ftc.teamcode.robot.CheckmateRobot
+import org.firstinspires.ftc.teamcode.robot.FreightFrenzyRobot
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode
@@ -29,7 +29,7 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode
  * Basic OpMode class that all OpModes should extend
  */
 abstract class BaseOpMode : LinearOpMode() {
-    protected lateinit var robot: CheckmateRobot
+    protected lateinit var robot: FreightFrenzyRobot
     protected lateinit var gp1: SuperController
     protected lateinit var gp2: SuperController
 
@@ -71,7 +71,7 @@ abstract class BaseOpMode : LinearOpMode() {
             opModeType = OpModeType.Autonomous
         }
         preSetup()
-        robot = CheckmateRobot(hardwareMap)
+        robot = FreightFrenzyRobot(hardwareMap)
         gp1 = SuperController(gamepad1)
         gp2 = SuperController(gamepad2)
         setup()
